Move EAC/EAD/EAG export functionality from admin to portal.
Not yet exposed/linked to on the interface (yet.)
Currently, whilst the URLs are uniform (/export), the formats for
each type respectively are hard-coded.
