+++ modules/admin/app/views/admin/documentaryUnit/adminActions.scala.html
@@ -26,6 +26,6 @@
         <a href="@controllers.admin.routes.ApiController.getItem(item.isA, item.id)">@Messages("item.export.json")</a>
     }
     @views.html.admin.common.sidebarAction() {
-        <a href="@controllers.units.routes.DocumentaryUnits.exportEad(item.topLevel.id)">@Messages("documentaryUnit.export.ead")</a>
+        <a href="@controllers.portal.routes.DocumentaryUnits.export(item.topLevel.id)">@Messages("documentaryUnit.export.ead")</a>
     }
 }
