+++ modules/admin/app/views/admin/repository/adminActions.scala.html
@@ -29,7 +29,7 @@
         <a href="@controllers.admin.routes.ApiController.getItem(item.isA, item.id)">@Messages("item.export.json")</a>
     }
     @views.html.admin.common.sidebarAction() {
-        <a href="@controllers.institutions.routes.Repositories.exportEag(item.id)">@Messages("repository.export.eag")
+        <a href="@controllers.portal.routes.Repositories.export(item.id)">@Messages("repository.export.eag")
         </a>
     }
 }
