+++ modules/admin/conf/institutions.routes
@@ -25,5 +25,4 @@ GET         /:id/link/:toType/:to                     @controllers.institutions.
 POST        /:id/link/:toType/:to                     @controllers.institutions.Repositories.linkAnnotatePost(id: String, toType: EntityType.Value, to: String)
 GET         /:id/reindex                              @controllers.institutions.Repositories.updateIndex(id: String)
 POST        /:id/reindex                              @controllers.institutions.Repositories.updateIndexPost(id: String)
-GET         /:id/eag                                  @controllers.institutions.Repositories.exportEag(id: String)
