+++ modules/portal/app/controllers/portal/HistoricalAgents.scala
@@ -6,10 +6,12 @@ import backend.rest.cypher.Cypher
 import com.google.inject.{Inject, Singleton}
 import controllers.generic.Search
 import controllers.portal.base.{Generic, PortalController}
+import defines.EntityType
 import models.HistoricalAgent
 import play.api.cache.CacheApi
 import play.api.i18n.MessagesApi
 import play.api.libs.concurrent.Execution.Implicits._
+import play.api.libs.iteratee.Enumerator
 import utils.MovedPageLookup
 import utils.search._
 import views.MarkdownRenderer
@@ -46,5 +48,15 @@ case class HistoricalAgents @Inject()(
 
   def browse(id: String) = GetItemAction(id).apply { implicit request =>
     Ok(views.html.historicalAgent.show(request.item, request.annotations, request.links, request.watched))
+  }
+
+  def export(id: String) = OptionalUserAction.async { implicit request =>
+    val format = "eac"
+    val params = request.queryString.filterKeys(_ == "lang")
+    userBackend.stream(s"${EntityType.HistoricalAgent}/$id/$format", params = params).map { case (head, body) =>
+      Status(head.status)
+        .chunked(body.andThen(Enumerator.eof))
+        .withHeaders(head.headers.map(s => (s._1, s._2.head)).toSeq: _*)
+    }
   }
 }
